duplicated functions are united
